Time: 54 ms (26.99%) | Memory: 16.5 MB (52.29%) - LeetSync

+++ 2571-find-the-pivot-integer/find-the-pivot-integer.py
@@ -15,8 +15,8 @@ class Solution:
         # print(i)
 
         for i in range(1,n+1):
-            a=int((i*(i+1))/2)
-            b=int(((n-i+1)/2)*(n+i))
+            a=(i*(i+1))/2
+            b=((n-i+1)/2)*(n+i)
             # print(a,b)
             if a==b:
                 return i
